Change to 100 microsecond delay for Art-Net
Seems like a 10 microsecond delay on boards like the ESP32-P4 may not be enough, causing some issues with UDP audio sync being delayed or appearing "buffered".  Oddly 10 microseconds was fine with the ESP32 and ESP32-S3, so it may also be an underlying IDF v5.3 change. 

Practically, 10 microseconds or 100 microseconds isn't a real difference when we're dealing in frames per second.